# openapi/bundled/openapi.yaml
openapi: 3.0.3
info:
  title: Wage App API
  version: 0.1.0
  description: Complete canonical contract for all authenticated /api/v1 Company, Employee, Shift Session, Checkpoint, Shift Result Draft, Shift Result, and Payroll operations currently provided by Wage App.
servers:
  - url: /
    description: Current Wage App server
security:
  - BearerAuth: []
...
paths:
  /api/v1/company/get/{id}:
    get:
      tags:
        - Company
      operationId: getCompany
      summary: Get a Company
      description: Returns one Company by its identifier.
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      responses:
        '200':
          description: The Company was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CompanyNotFound'
  /api/v1/company/get/for-user:
    get:
      tags:
        - Company
      operationId: getCompaniesForUser
      summary: List the current user's Companies
      description: Returns the Companies associated with the authenticated User's Employee record.
      responses:
        '200':
          description: The Companies associated with the current User, possibly empty.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCompaniesResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/company/get/page:
    get:
      tags:
        - Company
      operationId: getCompaniesPage
      summary: List a page of Companies
      description: Returns a page of Companies. The response guarantees the page content, index, requested size, and total element count; additional Spring pagination fields may be present but are not part of the supported contract.
      parameters:
        - $ref: '#/components/parameters/CompanyPageIndex'
        - $ref: '#/components/parameters/CompanyPageSize'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: A page of Companies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/company/create:
    post:
      tags:
        - Company
      operationId: createCompany
      summary: Create a Company
      description: Creates a Company with a unique title.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreateOrUpdateRequest'
      responses:
        '201':
          description: The Company was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/CompanyConflict'
  /api/v1/company/update/{id}:
    put:
      tags:
        - Company
      operationId: updateCompany
      summary: Update a Company
      description: Replaces the editable fields of an existing Company.
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreateOrUpdateRequest'
      responses:
        '204':
          description: The Company was updated successfully. The response has no body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/CompanyNotFound'
        '409':
          $ref: '#/components/responses/CompanyConflict'
  /api/v1/company/delete/{id}:
    delete:
      tags:
        - Company
      operationId: deleteCompany
      summary: Delete a Company
      description: Deletes a Company by identifier. Deleting an absent Company is also successful.
      parameters:
        - $ref: '#/components/parameters/CompanyId'
      responses:
        '204':
          description: The Company is absent after the request. The response has no body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/employee/get/{id}:
    get:
      tags:
        - Employee
      operationId: getEmployee
      summary: Get an Employee
      description: Returns one Employee, including all Company memberships and any User binding.
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '200':
          description: The Employee was found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/EmployeeNotFound'
  /api/v1/employee/get/all:
    get:
      tags:
        - Employee
      operationId: getAllEmployees
      summary: List all Employees
      description: Returns every Employee, including Company memberships and User bindings.
      responses:
        '200':
          description: All Employees, possibly empty.
          content:
            application/json:
              schema:
                type: array
                nullable: false
                items:
                  $ref: '#/components/schemas/Employee'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/employee/get/by-companies:
    get:
      tags:
        - Employee
      operationId: getEmployeesByCompanies
      summary: List Employees by Company
      description: Returns Employees grouped by each requested Company. A Company without a matching Employee is omitted rather than returned with an empty collection.
      parameters:
        - $ref: '#/components/parameters/CompanyIds'
      responses:
        '200':
          description: Company groups that contain at least one matching Employee, possibly empty.
          content:
            application/json:
              schema:
                type: array
                nullable: false
                items:
                  $ref: '#/components/schemas/CompanyEmployees'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/employee/get/coworkers:
    get:
      tags:
        - Employee
      operationId: getCoworkers
      summary: List the current User's coworkers
      description: Returns Employees, including the current User's Employee, grouped by each Company shared with the current User. Returns an empty list when the User is not bound to an Employee or that Employee has no Company membership.
      responses:
        '200':
          description: Company groups containing the current User's coworkers, possibly empty.
          content:
            application/json:
              schema:
                type: array
                nullable: false
                items:
                  $ref: '#/components/schemas/CompanyEmployees'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/employee/create:
    post:
      tags:
        - Employee
      operationId: createEmployee
      summary: Create an Employee
      description: Creates an Employee and establishes the requested Company memberships.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: The Employee was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/employee/update/{id}:
    put:
      tags:
        - Employee
      operationId: updateEmployee
      summary: Update an Employee
      description: Replaces the Employee's editable details and full Company membership list. The request can also replace or remove the User binding.
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '204':
          description: The Employee was updated successfully. The response has no body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/EmployeeNotFound'
        '409':
          $ref: '#/components/responses/UserBindingConflict'
  /api/v1/employee/bind-user/{employeeId}:
    put:
      tags:
        - Employee
      operationId: bindEmployeeUser
      summary: Bind an Employee to a User
      description: Replaces the Employee's User binding with the supplied external User identifier. Company memberships and Employee details are unchanged.
      parameters:
        - $ref: '#/components/parameters/EmployeeBindingId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: The User was bound to the Employee. The response has no body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/EmployeeNotFound'
        '409':
          $ref: '#/components/responses/UserBindingConflict'
  /api/v1/employee/delete/{id}:
    delete:
      tags:
        - Employee
      operationId: deleteEmployee
      summary: Delete an Employee
      description: Deletes an Employee and its Company memberships by identifier. Deleting an absent Employee is also successful.
      parameters:
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '204':
          description: The Employee is absent after the request. The response has no body.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A JWT access token accepted by the Wage App resource server.
  parameters:
    CompanyId:
      name: id
      in: path
      required: true
      description: The Company's identifier.
      schema:
        type: string
        format: uuid
        nullable: false
      example: 9b66d829-8909-4399-bbbe-8ad4b14b9251
    EmployeeId:
      name: id
      in: path
      required: true
      description: The Employee's identifier.
      schema:
        type: string
        format: uuid
        nullable: false
      example: 6a3c626f-0d0f-41d8-9818-06d263b25695
    EmployeeBindingId:
      name: employeeId
      in: path
      required: true
      description: The identifier of the Employee to bind to a User.
      schema:
        type: string
        format: uuid
        nullable: false
      example: 6a3c626f-0d0f-41d8-9818-06d263b25695
    CompanyIds:
      name: companyIds
      in: query
      required: true
      description: One or more Company identifiers supplied as repeated `companyIds` parameters. Only Companies with at least one matching Employee are represented in the response.
      style: form
      explode: true
      schema:
        type: array
        minItems: 1
        nullable: false
        items:
          type: string
          format: uuid
          nullable: false
      example:
        - 9b66d829-8909-4399-bbbe-8ad4b14b9251
    UserId:
      name: userId
      in: query
      required: true
      description: The external identifier of the User to bind. It must not already be bound to a different Employee.
      schema:
        type: string
        nullable: false
      example: keycloak-user-123
...
    Sort:
      name: sort
      in: query
      required: false
      description: Sort criterion in `property,direction` form. Repeat the parameter for multiple criteria. Direction is `asc` or `desc`; when omitted, ascending order is used.
      style: form
      explode: true
      schema:
        type: array
        nullable: false
        items:
          type: string
      example:
        - title,asc
    CompanyPageIndex:
      name: page
      in: query
      required: false
      description: Zero-based decimal page index. Missing, malformed, or negative values use the default of 0. This parameter is represented as a string so the generated boundary can preserve that lenient coercion behavior.
      schema:
        type: string
        default: '0'
        nullable: false
    CompanyPageSize:
      name: size
      in: query
      required: false
      description: Decimal number of Companies requested per page. Missing, malformed, or nonpositive values use the default of 100; values above 2000 are capped at 2000. This parameter is represented as a string so the generated boundary can preserve that lenient coercion behavior.
      schema:
        type: string
        default: '100'
        nullable: false
  responses:
    BadRequest:
      description: The path, query parameters, or request body are invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: A valid JWT was not supplied. The current response has no body.
    Forbidden:
      description: The authenticated principal is not allowed to perform the operation. The current response has no body.
    CompanyNotFound:
      description: The requested Company does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    CompanyConflict:
      description: Another Company already has the requested title.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    EmployeeNotFound:
      description: The requested Employee does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    UserBindingConflict:
      description: The requested User is already bound to a different Employee.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
    Company:
      type: object
      description: A work location whose Shift Sessions and Payroll are accounted independently.
      additionalProperties: false
      required:
        - id
        - title
        - employeeWageCoefficientFromRevenue
        - defaultShiftStartTime
      properties:
        id:
          type: string
          format: uuid
          nullable: false
          description: The Company's identifier.
        title:
          type: string
          minLength: 1
          maxLength: 255
          nullable: false
          description: The unique display title of the Company.
          example: Central Cafe
        employeeWageCoefficientFromRevenue:
          type: integer
          format: int32
          nullable: false
          description: The Company's Revenue Percentage expressed as a whole-number percentage.
          example: 15
        defaultShiftStartTime:
          type: string
          minLength: 5
          maxLength: 5
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
          nullable: false
          description: Default local Shift Session Start time in 24-hour `HH:mm` form.
          example: '09:00'
    CompanyCreateOrUpdateRequest:
      type: object
      description: Fields required to create or update a Company.
      additionalProperties: false
      required:
        - title
        - employeeWageCoefficientFromRevenue
        - defaultShiftStartTime
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          nullable: false
          description: The unique display title of the Company.
          example: Central Cafe
        employeeWageCoefficientFromRevenue:
          type: integer
          format: int32
          nullable: false
          description: The Company's Revenue Percentage expressed as a whole-number percentage.
          example: 15
        defaultShiftStartTime:
          type: string
          minLength: 5
          maxLength: 5
          pattern: ^([0-1][0-9]|2[0-3]):[0-5][0-9]$
          nullable: false
          description: Default local Shift Session Start time in 24-hour `HH:mm` form.
          example: '09:00'
    UserCompaniesResponse:
      type: object
      description: Companies associated with the authenticated User's Employee record.
      additionalProperties: false
      required:
        - companies
      properties:
        companies:
          type: array
          nullable: false
          items:
            $ref: '#/components/schemas/Company'
    CompanyPage:
      type: object
      description: Stable Company pagination fields. The server can emit additional framework metadata, which consumers must not rely on.
      additionalProperties: true
      required:
        - content
        - number
        - size
        - totalElements
      properties:
        content:
          type: array
          nullable: false
          description: Companies in this page.
          items:
            $ref: '#/components/schemas/Company'
        number:
          type: integer
          format: int32
          minimum: 0
          nullable: false
          description: Zero-based page index.
          example: 0
        size:
          type: integer
          format: int32
          minimum: 0
          nullable: false
          description: Requested page size, or zero for an unpaged response.
          example: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
          nullable: false
          description: Total number of Companies across all pages.
          example: 2
    Position:
      type: string
      description: The Employee's operational designation; it does not grant system permissions.
      nullable: false
      enum:
        - MANAGER
        - WAITER_ACTIVE
        - WAITER_INACTIVE
      example: WAITER_ACTIVE
    Employee:
      type: object
      description: A person whose work and pay are accounted by Companies. The current API also permits an Employee to exist without a Company membership.
      additionalProperties: false
      required:
        - id
        - companyIds
        - userId
        - firstName
        - lastName
        - patronymic
        - simpleName
        - position
      properties:
        id:
          type: string
          format: uuid
          nullable: false
          description: The Employee's identifier.
          example: 6a3c626f-0d0f-41d8-9818-06d263b25695
        companyIds:
          type: array
          nullable: false
          description: Identifiers of the Companies whose membership includes this Employee.
          items:
            type: string
            format: uuid
            nullable: false
          example:
            - 9b66d829-8909-4399-bbbe-8ad4b14b9251
        userId:
          type: string
          nullable: true
          description: The bound User's external identifier, or null when no User is bound.
          example: keycloak-user-123
        firstName:
          type: string
          nullable: false
          description: The Employee's first name.
          example: Ivan
        lastName:
          type: string
          nullable: false
          description: The Employee's family name.
          example: Ivanov
        patronymic:
          type: string
          nullable: false
          description: The Employee's patronymic.
          example: Ivanovich
        simpleName:
          type: string
          nullable: true
          description: An optional short display name for the Employee.
          example: Vanya
        position:
          $ref: '#/components/schemas/Position'
    CompanyEmployee:
      type: object
      description: Employee details returned within one Company grouping.
      additionalProperties: false
      required:
        - id
        - userId
        - firstName
        - lastName
        - patronymic
        - simpleName
        - position
      properties:
        id:
          type: string
          format: uuid
          nullable: false
          description: The Employee's identifier.
          example: 6a3c626f-0d0f-41d8-9818-06d263b25695
        userId:
          type: string
          nullable: true
          description: The bound User's external identifier, or null when no User is bound.
          example: keycloak-user-123
        firstName:
          type: string
          nullable: false
          description: The Employee's first name.
          example: Ivan
        lastName:
          type: string
          nullable: false
          description: The Employee's family name.
          example: Ivanov
        patronymic:
          type: string
          nullable: false
          description: The Employee's patronymic.
          example: Ivanovich
        simpleName:
          type: string
          nullable: true
          description: An optional short display name for the Employee.
          example: Vanya
        position:
          $ref: '#/components/schemas/Position'
    CompanyEmployees:
      type: object
      description: Employees whose membership includes one Company.
      additionalProperties: false
      required:
        - companyId
        - data
      properties:
        companyId:
          type: string
          format: uuid
          nullable: false
          description: The Company's identifier.
          example: 9b66d829-8909-4399-bbbe-8ad4b14b9251
        data:
          type: array
          nullable: false
          minItems: 1
          description: Employees belonging to this Company.
          items:
            $ref: '#/components/schemas/CompanyEmployee'
    CreateEmployeeRequest:
      type: object
      description: Fields used to create an Employee and establish initial Company memberships.
      additionalProperties: false
      required:
        - firstName
        - lastName
        - patronymic
        - position
      properties:
        companyIds:
          type: array
          nullable: false
          default: []
          description: Identifiers of the Companies the new Employee belongs to. Omission creates the Employee without Company membership.
          items:
            type: string
            format: uuid
            nullable: false
          example:
            - 9b66d829-8909-4399-bbbe-8ad4b14b9251
        firstName:
          type: string
          minLength: 1
          pattern: .*\S.*
          nullable: false
          description: The Employee's first name; it must contain a non-whitespace character.
          example: Ivan
        lastName:
          type: string
          minLength: 1
          pattern: .*\S.*
          nullable: false
          description: The Employee's family name; it must contain a non-whitespace character.
          example: Ivanov
        patronymic:
          type: string
          minLength: 1
          pattern: .*\S.*
          nullable: false
          description: The Employee's patronymic; it must contain a non-whitespace character.
          example: Ivanovich
        simpleName:
          type: string
          nullable: true
          description: An optional short display name for the Employee.
          example: Vanya
        position:
          $ref: '#/components/schemas/Position'
    UpdateEmployeeRequest:
      type: object
      description: Fields used to replace an Employee's editable details and complete Company membership list. A null or omitted `userId` removes any existing User binding.
      additionalProperties: false
      required:
        - firstName
        - lastName
        - patronymic
        - position
      properties:
        companyIds:
          type: array
          nullable: false
          default: []
          description: Complete replacement list of Company memberships. Omission removes all Company memberships.
          items:
            type: string
            format: uuid
            nullable: false
          example:
            - 9b66d829-8909-4399-bbbe-8ad4b14b9251
        userId:
          type: string
          nullable: true
          description: External identifier of the User to bind, or null/omitted to remove the current binding. A non-null value must be unique across Employees.
          example: keycloak-user-123
        firstName:
          type: string
          minLength: 1
          pattern: .*\S.*
          nullable: false
          description: The Employee's first name; it must contain a non-whitespace character.
          example: Ivan
        lastName:
          type: string
          minLength: 1
          pattern: .*\S.*
          nullable: false
          description: The Employee's family name; it must contain a non-whitespace character.
          example: Ivanov
        patronymic:
          type: string
          minLength: 1
          pattern: .*\S.*
          nullable: false
          description: The Employee's patronymic; it must contain a non-whitespace character.
          example: Ivanovich
        simpleName:
          type: string
          nullable: true
          description: An optional short display name for the Employee.
          example: Vanya
        position:
          $ref: '#/components/schemas/Position'
    ProblemDetail:
      type: object
      description: RFC 7807 problem details returned for application-level errors.
      additionalProperties: true
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri-reference
          nullable: false
          description: Identifier for the problem type; currently `about:blank`.
          example: about:blank
        title:
          type: string
          nullable: false
          description: Short HTTP status phrase.
          example: Not Found
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
          nullable: false
          description: HTTP status code.
          example: 404
        detail:
          type: string
          nullable: false
          description: Human-readable explanation specific to this occurrence.
          example: Company 5b367932-2e5d-4d89-aa07-17f65ff4af9c not found
        instance:
          type: string
          format: uri-reference
          nullable: false
          description: Request path associated with this occurrence when supplied by the server.
          example: /api/v1/company/get/5b367932-2e5d-4d89-aa07-17f65ff4af9c
...